refactor: quality-audit fixes — classify-then-consume retries, retry.go split, report.Aggregate
From the code-quality audit:

- getPkgInformation: the switch's len-guards let an exhausted class bleed
  into the transient budget (a persistent dial failure got dial+transient
  sleeps), contradicting the stated contract. Restructure to classify the
  error first, then consume from exactly that class's schedule; empty
  schedule = return for every class. Sleeps are now ctx-aware (sleepCtx),
  as is the env-specialize wait.
- Split retry policy (schedules, wait budget, delay fn, sleepCtx) into
  pkg/fetcher/retry.go with its tests — fetcher.go was parked at exactly
  1000 lines and the policy/mechanics seam is natural.
- addForwardedHostHeader: remove the bracket-strip else-branch that
  existed to feed the deleted ParseIP — provably dead for the quoting
  test (bracketed hosts match via brackets/colons either way; pinned by
  the existing bracketed-port-less test).
- Move aggregateReps to report.Aggregate: median-folding ScenarioResults
  is pure report-model computation, which that package documents as its
  boundary; tests move along.
- coldburst: strconv.Itoa for int meta, matching sibling files.
